+++ tests/e2e/connector_modify_test.go
@@ -44,7 +44,7 @@ func TestNewConnectorModifyE2E(t *testing.T) {
 	testutils.AssertEqual(t, *updated.Data.SyncFrequency, 1440)
 	testutils.AssertEqual(t, updated.Data.ScheduleType, "auto")
 	testutils.AssertEmpty(t, updated.Data.PrivateLinkId)
-	testutils.AssertEmpty(t, updated.Data.LocalProcessingAgentId)
+	testutils.AssertEmpty(t, updated.Data.HybridDeploymentAgentId)
 	testutils.AssertEmpty(t, updated.Data.ProxyAgentId)
 	testutils.AssertEqual(t, updated.Data.NetworkingMethod, "Directly")