+++ CHANGELOG.md
@@ -5,7 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/fivetran/go-fivetran/compare/v0.9.4...HEAD)
+## [Unreleased](https://github.com/fivetran/go-fivetran/compare/v1.0.0...HEAD)
+
+## [1.0.0](https://github.com/fivetran/go-fivetran/compare/v0.9.4...v1.0.0)
+
+Release version
+
+Renamed the following service from Local Processing Agent to Hybrid Deployment Agent:
+- LocalProcessingAgentCreateService to HybridDeploymentAgentCreateService
+- LocalProcessingAgentListService to HybridDeploymentAgentListService
+- LocalProcessingAgentDetailsService to HybridDeploymentAgentDetailsService
+- LocalProcessingAgentDeleteService to HybridDeploymentAgentDeleteService
+- LocalProcessingAgentReAuthService to HybridDeploymentAgentReAuthService
+
+Rename fields:
+- `DestinationCreateService.localProcessingAgentId` to `DestinationCreateService.hybridDeploymentAgentId`
+- `DestinationModifyService.localProcessingAgentId` to `DestinationCreateService.hybridDeploymentAgentId`
+- `DestinationDetailsService.localProcessingAgentId` to `DestinationCreateService.hybridDeploymentAgentId`
+- `ConnectorCreateService.localProcessingAgentId` to `DestinationCreateService.hybridDeploymentAgentId`
+- `ConnectorModifyService.localProcessingAgentId` to `DestinationCreateService.hybridDeploymentAgentId`
+- `ConnectorDetailsService.localProcessingAgentId` to `DestinationCreateService.hybridDeploymentAgentId`
 
 ## [0.9.3](https://github.com/fivetran/go-fivetran/compare/v0.9.2...v0.9.4)